+++ OnlineShop/Controllers/Api/DishManagementSystemApiController.cs
@@ -22,35 +22,65 @@ public class DishManagementSystemApiController : ControllerBase
     [HttpGet(Routes.All)]
     public async Task<IActionResult> GetAllDishesAsync(CancellationToken token)
     {
-        var dishes = await _dishRepository.GetAllAsync(token);
-        var viewModel = _mapper.Map<List<DishModel>>(dishes);
-        
-        return Ok(viewModel);
+        try
+        {
+            var dishes = await _dishRepository.GetAllAsync(token);
+            var viewModel = _mapper.Map<List<DishModel>>(dishes);
+
+            return Ok(viewModel);
+        }
+        catch
+        {
+            return NotFound("Failed to get all dishes");
+        }
     }
 
     [HttpPost(Routes.Create)]
-    public async Task<IActionResult> CreateDishAsync(DishModel dish, CancellationToken token)
+    public async Task<IActionResult> CreateDishAsync([FromBody] DishModel dish, CancellationToken token)
     {
-        var data = _mapper.Map<Dish>(dish);
-        await _dishRepository.CreateAsync(data, token);
-        
-        return Ok();
-    }
-    
-    [HttpPost(Routes.Update)]
-    public async Task<IActionResult> UpdateDishAsync(DishModel dish, CancellationToken token)
-    {
-        var data = _mapper.Map<Dish>(dish);
-        await _dishRepository.UpdateAsync(data, token);
-        
-        return Ok();
-    }
-    
-    [HttpPost(Routes.Delete)]
-    public async Task<IActionResult> DeleteDishAsync(Guid id, CancellationToken token)
-    {
-        await _dishRepository.DeleteAsync(id, token);
+        try
+        {
+            var data = _mapper.Map<Dish>(dish);
+            await _dishRepository.CreateAsync(data, token);
 
-        return Ok();
+            return Ok();
+        }
+        catch
+        {
+            return BadRequest("Failed to create a dish");
+        }
+    }
+
+    [HttpPost(Routes.Update)]
+    public async Task<IActionResult> UpdateDishAsync([FromBody] DishModel dish, CancellationToken token)
+    {
+        try
+        {
+            var data = _mapper.Map<Dish>(dish);
+            await _dishRepository.UpdateAsync(data, token);
+
+            return Ok();
+        }
+        catch
+        {
+            return BadRequest("Failed to update dish");
+        }
+    }
+
+    [HttpPost(Routes.Delete)]
+    public async Task<IActionResult> DeleteDishAsync([FromQuery] Guid id, CancellationToken token)
+    {
+        try
+        {
+            if (id == Guid.Empty)
+                return BadRequest("Incorrect Id specified");
+            
+            await _dishRepository.DeleteAsync(id, token);
+            return Ok();
+        }
+        catch
+        {
+            return BadRequest("Failed to delete dish");
+        }
     }
 }
